feat(Worker Pool): Creates new `fromWorkerPool` method to automatically  manage pools of workers for parallel processing

+++ projects/observable-webworker/src/lib/from-worker.ts
@@ -2,10 +2,15 @@ import { Observable, Observer, Subscription, Notification } from 'rxjs';
 import { dematerialize, map, materialize, tap } from 'rxjs/operators';
 import { GenericWorkerMessage, WorkerMessageNotification } from './observable-worker.types';
 
+export interface WorkerOptions {
+  terminateOnComplete: boolean;
+}
+
 export function fromWorker<Input, Output>(
   workerFactory: () => Worker,
   input$: Observable<Input>,
   selectTransferables?: (input: Input) => Transferable[],
+  options: WorkerOptions = { terminateOnComplete: true },
 ): Observable<Output> {
   return new Observable((responseObserver: Observer<Notification<GenericWorkerMessage<Output>>>) => {
     let worker: Worker;
@@ -39,7 +44,7 @@ export function fromWorker<Input, Output>(
       if (subscription) {
         subscription.unsubscribe();
       }
-      if (worker) {
+      if (worker && options.terminateOnComplete) {
         worker.terminate();
       }
     };
